common: move the privilege dropping code into a dedicated file.
The intention is to share it between crimson and the classical OSD.
